Move cf/korifi NS creation into terraform templates
We now apply the [env]/k8s terraform before deploying korifi,
and destroy it after deployment (with a sneaky deletion of just the cf
namespace before korifi deletion)

+++ pipelines/scripts/delete-cluster.sh
@@ -14,7 +14,7 @@ pushd cf-k8s-secrets/ci-deployment/$CLUSTER_NAME || exit 1
     -backend-config="prefix=terraform/state/$CLUSTER_NAME"
 
   cluster_values="$(terraform show -json | jq -r '.values ')"
-  if [ "$cluster_values" == "null" ]; then
+  if [[ "$cluster_values" == "null" ]]; then
     echo "Cluster $CLUSTER_NAME does not exist"
     exit 0
   fi
@@ -22,48 +22,30 @@ pushd cf-k8s-secrets/ci-deployment/$CLUSTER_NAME || exit 1
   export-kubeconfig
   kubectl get service --all-namespaces | awk '/LoadBalancer/ { print "kubectl delete service -n " $1 " " $2 }' | xargs -IN sh -c "N"
 
-  case "${CLUSTER_TYPE:-}" in
-    "GKE")
-      cluster_network="$(terraform show -json | jq -r '.values.root_module.resources[] | select(.name == "network") | .values.name')"
-      if [[ -n "$cluster_network" ]]; then
-        gcloud-login
+  if [[ "${CLUSTER_TYPE:-}" == "GKE" ]]; then
+    cluster_network="$(terraform show -json | jq -r '.values.root_module.resources[] | select(.name == "network") | .values.name')"
+    if [[ -n "$cluster_network" ]]; then
+      gcloud-login
 
-        # Firewall rules are created when a LoadBalancer Service is created in GCP,
-        # and since terraform doesn't know about them, they have to be manually deleted.
-        # See https://github.com/terraform-providers/terraform-provider-google/issues/5948
-        firewall_rules=$(gcloud compute firewall-rules list --filter=network="$cluster_network" --format="value(name)")
-        for rule_name in $firewall_rules; do
-          gcloud compute firewall-rules delete "$rule_name" --quiet || echo "firewall rule $rule_name not found"
-        done
+      # Firewall rules are created when a LoadBalancer Service is created in GCP,
+      # and since terraform doesn't know about them, they have to be manually deleted.
+      # See https://github.com/terraform-providers/terraform-provider-google/issues/5948
+      firewall_rules=$(gcloud compute firewall-rules list --filter=network="$cluster_network" --format="value(name)")
+      for rule_name in $firewall_rules; do
+        gcloud compute firewall-rules delete "$rule_name" --quiet || echo "firewall rule $rule_name not found"
+      done
 
-        # Network endpoint groups are created for the rapid chanel version of Kubernetes
-        # clusters on GCP (couldn't find a docs link). Since terraform doesn't know about
-        # these resources they also must be deleted separately.
-        network_endpoint_groups_csv=$(gcloud beta compute network-endpoint-groups list --filter="network ~ $CLUSTER_NAME" --format="csv[no-heading](name,zone)")
-        for name_zone in $network_endpoint_groups_csv; do
-          name="$(echo "$name_zone" | awk -F ',' '{ print $1 }')"
-          zone="$(echo "$name_zone" | awk -F ',' '{ print $2 }')"
-          gcloud beta compute network-endpoint-groups delete --zone "$zone" "$name" --quiet
-        done
-      fi
-      ;;
-
-    "EKS")
-      pushd k8s || exit 1
-      {
-        terraform init \
-          -backend-config="prefix=terraform/state/$CLUSTER_NAME-k8s"
-
-        terraform destroy \
-          -var "name=$CLUSTER_NAME" \
-          -auto-approve
-      }
-      popd || exit 1
-      ;;
-  esac
-
-  terraform init \
-    -backend-config="prefix=terraform/state/$CLUSTER_NAME"
+      # Network endpoint groups are created for the rapid chanel version of Kubernetes
+      # clusters on GCP (couldn't find a docs link). Since terraform doesn't know about
+      # these resources they also must be deleted separately.
+      network_endpoint_groups_csv=$(gcloud beta compute network-endpoint-groups list --filter="network ~ $CLUSTER_NAME" --format="csv[no-heading](name,zone)")
+      for name_zone in $network_endpoint_groups_csv; do
+        name="$(echo "$name_zone" | awk -F ',' '{ print $1 }')"
+        zone="$(echo "$name_zone" | awk -F ',' '{ print $2 }')"
+        gcloud beta compute network-endpoint-groups delete --zone "$zone" "$name" --quiet
+      done
+    fi
+  fi
 
   terraform destroy \
     -var "name=$CLUSTER_NAME" \
